--- packages/wranglerjs-compat-webpack-plugin/src/__tests__/helpers/mock-webpack-plugin.ts
@@ -1,7 +1,0 @@
-import { execa } from "execa";
-
-export const buildWebpackPlugin = () => {
-  beforeAll(async () => {
-    await execa("npm", ["run", "build:js"]);
-  });
-};
